+++ great_expectations/experimental/datasources/__init__.py
@@ -1,3 +1,6 @@
+from great_expectations.experimental.datasources.pandas_datasource import (
+    PandasDatasource,
+)
 from great_expectations.experimental.datasources.postgres_datasource import (
     PostgresDatasource,
 )
