+++ great_expectations/experimental/datasources/sources.py
@@ -9,7 +9,7 @@ from great_expectations.experimental.datasources.type_lookup import TypeLookup
 
 if TYPE_CHECKING:
     from great_expectations.data_context import AbstractDataContext as GXDataContext
-    from great_expectations.experimental.datasources.context import DataContext
+    from great_expectations.experimental.context import DataContext
     from great_expectations.experimental.datasources.interfaces import (
         DataAsset,
         Datasource,
